feat: cell diff options
Problem: setting escape sequences, e.g. image protocols, always causes a
cell to count as "multiwidth", even when it doesn't render wider than
one cell. For images specifically, when rendering a tiny image into just
one cell, the cell to the right will always be skipped - even if that
cell has nothing to do with the widget / buffer that the image is
operating on.
In general, this affects any escape sequence or otherwise invisible text
in a symbol that does not actually render wider than one cell: the next
cell will be erroneously skipped.

Solution: Convert `skip:bool` to enum. Add enum option `UnitWidth` to
force a cell width of 1 for diffing.

When using the option, this also fixes some bug where diffing is not
idempotent and causes a diff operation for `(symbol.len() - 1)` times.